+++ examples/resources/bunnynet_dns_record/import.sh
@@ -1,0 +1,1 @@
+terraform import bunnynet_dns_record.test "$ZONE_ID|$RECORD_ID"