+++ examples/resources/bunnynet_dns_zone/import.sh
@@ -1,0 +1,1 @@
+terraform import bunnynet_dns_zone.test 123