+++ examples/resources/bunnynet_pullzone/import.sh
@@ -1,0 +1,1 @@
+terraform import bunnynet_pullzone.test 123