+++ examples/resources/bunnynet_pullzone_edgerule/import.sh
@@ -1,0 +1,1 @@
+terraform import bunnynet_pullzone_edgerule.test "$PULLZONE_ID|$EDGERULE_GUID"