+++ examples/resources/bunnynet_pullzone_hostname/import.sh
@@ -1,0 +1,1 @@
+terraform import bunnynet_pullzone_hostname.test "$PULLZONE_ID|example.com"