+++ examples/resources/bunnynet_pullzone_optimizer_class/import.sh
@@ -1,0 +1,1 @@
+terraform import bunnynet_pullzone_optimizer_class.test "$PULLZONE_ID|thumbnail"