+++ examples/resources/bunnynet_storage_file/import.sh
@@ -1,0 +1,1 @@
+terraform import bunnynet_storage_file.test "$STORAGE_ZONE_ID|blog/index.html"