+++ examples/resources/bunnynet_storage_zone/import.sh
@@ -1,0 +1,1 @@
+terraform import bunnynet_storage_zone.test 123