+++ examples/resources/bunnynet_stream_collection/import.sh
@@ -1,0 +1,1 @@
+terraform import bunnynet_stream_collection.test "$LIBRARY_ID|$COLLECTION_GUID"