+++ examples/resources/bunnynet_stream_library/import.sh
@@ -1,0 +1,1 @@
+terraform import bunnynet_stream_library.test 123