+++ examples/resources/bunnynet_stream_video/import.sh
@@ -1,0 +1,1 @@
+terraform import bunnynet_stream_video.test "$LIBRARY_ID|$VIDEO_GUID"